requirements.txt
Pillow>=10.0.0
pillow-heif>=0.18.0
pyoxipng>=9.0.0
tkinterdnd2>=0.3.0
